+++ R/constants.R
@@ -58,29 +58,31 @@ CONSTANTS <- list(
 CONSTANTS$PLANTS_PER_PLOT <- CONSTANTS$PLANTS_IN_X * CONSTANTS$PLANTS_IN_Y
 
 # Thornley Transport Resistance (TTR) constants (mirroring C++ reference)
-CONSTANTS$K_LITTER      <- 0.05
-CONSTANTS$K_M_LITTER    <- 2.5
-CONSTANTS$G_SHOOT       <- 200
-CONSTANTS$G_ROOT        <- 200
-CONSTANTS$G_DEFENCE     <- 10
-CONSTANTS$K_C           <- 0.1
-CONSTANTS$K_N           <- 0.01
-CONSTANTS$K_M           <- 10
-CONSTANTS$PI_C          <- 0.1
-CONSTANTS$PI_N          <- 0.01
-CONSTANTS$Q_SCP         <- 2/3
-CONSTANTS$TR_C          <- 1.0
-CONSTANTS$TR_N          <- 1.0
-CONSTANTS$FRACTION_C    <- 0.5
-CONSTANTS$FRACTION_N    <- 0.025
-CONSTANTS$PHENO_SWITCH  <- 10
-CONSTANTS$ACCEL_LEAF_LOSS <- 10
-CONSTANTS$TEMP_GROWTH_1 <- 1
-CONSTANTS$TEMP_GROWTH_2 <- 24
-CONSTANTS$TEMP_GROWTH_3 <- 26
-CONSTANTS$TEMP_GROWTH_4 <- 40
-CONSTANTS$TEMP_PHOTO_1  <- 1
-CONSTANTS$TEMP_PHOTO_2  <- 15
-CONSTANTS$TEMP_PHOTO_3  <- 25
-CONSTANTS$TEMP_PHOTO_4  <- 35
-CONSTANTS$HERBIVORY     <- 1
+CONSTANTS$K_LITTER        <- 0.05   # per day: litter production rate (proportion/day)
+CONSTANTS$K_M_LITTER      <- 2.5    # kg XDM: litter mass scale parameter
+CONSTANTS$G_SHOOT         <- 200    # per day scaling: shoot growth [kg C/kg N/kg XDM^2]/day
+CONSTANTS$G_ROOT          <- 200    # per day scaling: root growth [kg C/kg N/kg XDM^2]/day
+CONSTANTS$G_DEFENCE       <- 10     # per day scaling: defence growth [kg C/kg N/kg XDM^2]/day
+CONSTANTS$K_C             <- 0.1    # per day: C input rate [kg C/kg shoot XDM/day]
+CONSTANTS$K_N             <- 0.01   # per day: N input rate [kg N/kg shoot XDM/day]
+CONSTANTS$K_M             <- 10     # kg XDM: mass scale for uptake saturation
+CONSTANTS$PI_C            <- 0.1    # kg C/kg XDM: product inhibition for C uptake
+CONSTANTS$PI_N            <- 0.01   # kg N/kg XDM: product inhibition for N uptake
+CONSTANTS$Q_SCP           <- 2/3    # unitless: transport resistance exponent
+CONSTANTS$TR_C            <- 1.0    # XDM^(Q_SCP-1)/day: C transport resistance scale
+CONSTANTS$TR_N            <- 1.0    # XDM^(Q_SCP-1)/day: N transport resistance scale
+CONSTANTS$FRACTION_C      <- 0.5    # kg C/kg XDM: carbon fraction of dry matter
+CONSTANTS$FRACTION_N      <- 0.025  # kg N/kg XDM: nitrogen fraction of dry matter
+CONSTANTS$PHENO_SWITCH    <- 10     # deg C: phenology temperature threshold
+CONSTANTS$ACCEL_LEAF_LOSS <- 10     # unitless multiplier: accelerated leaf loss factor
+CONSTANTS$TEMP_GROWTH_1   <- 1      # deg C: growth envelope parameter
+CONSTANTS$TEMP_GROWTH_2   <- 24     # deg C: growth envelope parameter
+CONSTANTS$TEMP_GROWTH_3   <- 26     # deg C: growth envelope parameter
+CONSTANTS$TEMP_GROWTH_4   <- 40     # deg C: growth envelope parameter
+CONSTANTS$TEMP_PHOTO_1    <- 1      # deg C: photosynthesis envelope parameter
+CONSTANTS$TEMP_PHOTO_2    <- 15     # deg C: photosynthesis envelope parameter
+CONSTANTS$TEMP_PHOTO_3    <- 25     # deg C: photosynthesis envelope parameter
+CONSTANTS$TEMP_PHOTO_4    <- 35     # deg C: photosynthesis envelope parameter
+CONSTANTS$INIT_SW         <- 500    # L: initial standing water (for scenarios)
+CONSTANTS$INIT_N          <- 3.75   # kg N: initial soil N availability (for scenarios)
+CONSTANTS$HERBIVORY       <- 1      # 0/1 switch: enable herbivory in daily loop